Changes Made:
Task.java

	line 14: startDate changed to long
	line 16 endDate changed to long
	line 120 changed return value of getEndDate() to long
	line 127 changed return value of getStartDate() to long
DataBase.java

	line:98 changed startDate and endDate to be long instead of int

+++ app/src/main/java/com/uottawa/choremanager/DataBase.java
@@ -94,7 +94,8 @@ public class DataBase extends Application{
     }
 
     //Same purpose as above, except for tasks
-    public Task addTask(String name, int startDate, String description, int endDate, String ownerId, ArrayList<SubTask> materials, String status){
+    //Changed int to long
+    public Task addTask(String name, long startDate, String description, long endDate, String ownerId, ArrayList<SubTask> materials, String status){
         Task toAdd = new Task(name, startDate, description, endDate, ownerId, status);
 
         String id = dbTasks.push().getKey();
